+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 
 ### Features
 
-- Add retry when entry is locked. 
+- Add retry when entry is locked. Defaults to 30 seconds. Allows setting by `EntriesClient.RetryIfLockedForTimeout`
 
 ## 2.0.1